backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.2.1
python-multipart>=0.0.18
aiofiles>=24.1.0
boto3>=1.35.0
mangum>=0.19.0
httpx>=0.28.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
